chore(framework): change header to H4 in modal and toolbox
change header to H4 in modal and toolbox

Closes: DEV-10299
